LayoutTests/platform/wpe/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/css-module/content-type-checking-expected.txt

CONSOLE MESSAGE: TypeError: 'text/css' is not a valid JavaScript MIME type.
CONSOLE MESSAGE: TypeError: 'application/css' is not a valid JavaScript MIME type.
CONSOLE MESSAGE: TypeError: 'text/plain' is not a valid JavaScript MIME type.
CONSOLE MESSAGE: TypeError: 'text/css' is not a valid JavaScript MIME type.
CONSOLE MESSAGE: TypeError: 'text/css' is not a valid JavaScript MIME type.

FAIL text/css assert_unreached: Reached unreachable code
PASS application/css
PASS text/html+css
FAIL text/css;boundary=something assert_unreached: Reached unreachable code
FAIL text/css;foo=bar assert_unreached: Reached unreachable code
